rnf-join-fq => rnftools merge
