+++ setup.cfg
@@ -29,9 +29,6 @@ packages =
 	rnftools.mishmash
 	rnftools.lavender
 	rnftools.rnfformat
-scripts = 
-	bin/rnf-join-fq.py
-	#bin/rnftools.py
 
 extra_files = 
 	setup.py
